+++ lib/common.h
@@ -17,8 +17,8 @@
 
 char* common_strdup(char* input);
 int common_rand(void* target, size_t bytes);
-int common_strrepl(char* buffer, unsigned length, char* variable, char* replacement);
-char* common_strappf(char* target, unsigned* target_allocated, char* fmt, ...);
+int common_strrepl(char* buffer, size_t length, char* variable, char* replacement);
+char* common_strappf(char* target, size_t* target_allocated, char* fmt, ...);
 int common_tprintf(char* format, time_t time, char* buffer, size_t buffer_length);
 ssize_t common_read_file(char* filename, uint8_t** out);
 ssize_t common_next_line(char* buffer, size_t* append_offset_p, ssize_t* new_bytes_p);
